requirements.txt
pandas>=1.3
numpy>=1.21
kafka-python>=2.0
